% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_SPE_shares.R
\name{calculate_SPE_share}
\alias{calculate_SPE_share}
\title{calculate SPE shares}
\usage{
calculate_SPE_share(selected_country, years)
}
\arguments{
\item{selected_country}{the country to be analyzed}

\item{years}{the start and end year for which data should be downloaded}
}
\value{
a dataset with the SPE shares from OECD countries where data exists
}
\description{
calculate SPE shares
}
