% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_and_indicators.R
\name{calculate_weights}
\alias{calculate_weights}
\title{calculate weights}
\usage{
calculate_weights(data)
}
\arguments{
\item{data}{is the standard dataframe}
}
\value{
the dataframe containing the different weights
}
\description{
calculate weights
}
