% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_cdis.R
\name{clean_cdis}
\alias{clean_cdis}
\title{clean cdis}
\usage{
clean_cdis(cdis)
}
\arguments{
\item{cdis}{is the cdis dataset}
}
\value{
the cleaned cdis dataframe
}
\description{
function for basic cleaning of cdis data frame: removing of values for aggregate geographical zones, adding countrycodes, pivot longer, and express values in Millions
}
