% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_deloitte_wth.R
\name{clean_deloitte_wth}
\alias{clean_deloitte_wth}
\title{Clean Deloitte Withholding tax table}
\usage{
clean_deloitte_wth(tableslist)
}
\arguments{
\item{tableslist}{the not yet cleaned table}
}
\value{
the cleaned table
}
\description{
Clean Deloitte Withholding tax table
}
\examples{
#
}
