% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_original.R
\name{clean_original}
\alias{clean_original}
\title{clean original report}
\usage{
clean_original(df)
}
\arguments{
\item{df}{is a dataframe}
}
\value{
the cleaned dataframe
}
\description{
clean original report
}
