% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdi_debt_equity_graph.R
\name{debt_equity_graph}
\alias{debt_equity_graph}
\title{FDI debt equity graph}
\usage{
debt_equity_graph(df, taxhavens)
}
\arguments{
\item{df}{is the fdi data table, assumed to be CDIS}

\item{taxhavens}{is a list of tax havens}
}
\value{
a graph with debt equity ratios for inward and outward investment
}
\description{
FDI debt equity graph
}
