% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_datasets.R
\name{download_dataset}
\alias{download_dataset}
\title{download datasets through APIs}
\usage{
download_dataset(all_years)
}
\arguments{
\item{all_years}{is a vector of years for which the datasets should be downloaded}
}
\description{
download datasets through APIs
}
