% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_deloitte_wth.R
\name{extract_deloitte_wth}
\alias{extract_deloitte_wth}
\title{Extract Deloitte Withholding tax table}
\usage{
extract_deloitte_wth(file)
}
\arguments{
\item{file}{a path to a file}
}
\value{
the extracted, not yet cleaned table
}
\description{
Extract Deloitte Withholding tax table
}
