% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_maatm.R
\name{extract_maatm}
\alias{extract_maatm}
\title{extracting signatories of Multilateral Convention on Mutual Administrative Assistance in Tax Matters}
\usage{
extract_maatm(file)
}
\arguments{
\item{file}{is a pdf file}
}
\value{
the cleaned dataframe
}
\description{
extracting signatories of Multilateral Convention on Mutual Administrative Assistance in Tax Matters
}
