% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_action5_tables.R
\name{extract_tables_region}
\alias{extract_tables_region}
\title{extract tables with python camelot using a table region}
\usage{
extract_tables_region(reportparameterlist)
}
\arguments{
\item{reportparameterlist}{is a list of parameters containing filename, filepath and pagenumbers, and maybe region of the tables to be extracted}

\item{regionspec}{specifies whether the table should be extract with a specified}
}
\value{
a list of the extracted tables
}
\description{
extract tables with python camelot using a table region
}
