% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_pdf_tbl2.R
\name{t_pdf_tbl2}
\alias{t_pdf_tbl2}
\title{transform table output2}
\usage{
t_pdf_tbl2(df, pl, column)
}
\arguments{
\item{df}{is the dataframe to be transformed}

\item{pl}{is a vector containing the numbers of the "principle lines". Can for example be obtained by a formula like the following: which(is.na(df[,1]), arr.ind=TRUE)}

\item{column}{is the column to be copied}
}
\value{
returns the transformed dataframe
}
\description{
When, for example, exporting a table from a pdf, then content belonging to one line is often scattered across several lines.
This function copies lines that belong together in one and deletes the rest.
}
