% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxhavenshare.R
\name{taxhavenshare}
\alias{taxhavenshare}
\title{Calculating tax haven share}
\usage{
taxhavenshare(df, taxhavens, spe_table = NULL, highlightcountry = NULL)
}
\arguments{
\item{df}{a dataframe of}
}
\description{
Calculating tax haven share
}
