% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treaties_in_year.R
\name{treaties_in_year}
\alias{treaties_in_year}
\title{treaties in year}
\usage{
treaties_in_year(year, data)
}
\arguments{
\item{year}{is the year for which the network should be reconstructed}

\item{data}{is the tax treaty data, needs columns "ended" and "EFFECTIVE"}
}
\value{
returns the dataset including new column with the datayear
}
\description{
showing which treaties are in force in a giving year
}
